refactor: introduce application runtime crate
